app/requirements.txt
# requirements.txt

Flask
python-dotenv